requirements.txt
beautifulsoup4==4.11.2
boto3==1.26.97
celery==5.3.6
channels==4.1.0
coverage==6.5.0
daphne==4.2.1
dj-database-url==2.3.0
django==5.2.7
django-apscheduler==0.7.0
django-celery-results==2.4.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==24.3
django-jazzmin==3.0.1
django-jsonform==2.15.0
django-json-widget==2.0.1
django-redis==6.0.0
django-rest-passwordreset==1.3.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.4
idna==3.10
iniconfig==2.1.0
mnemonic==0.21
packaging==25.0
pillow==10.0.0
pluggy==1.6.0
postmarker==1.0
psycopg[binary]>=3.2
pygments==2.19.2
pyjwt==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-json-logger==4.0.0
pyyaml==6.0.3
redis==6.4.0
ruff==0.14.0
sentry-sdk==2.7
sniffio==1.3.1
sqlparse==0.5.3
sshkey-tools==0.10.1
typing-extensions==4.15.0 ; python_full_version < '3.13'
tzdata==2025.2 ; sys_platform == 'win32'
unidecode==1.3.7
urllib3==1.26.11
web3==6.20.2
whitenoise==6.3.0
python-jose[cryptography]>=3.3.0,<4
django-prometheus==2.4.1
